organizing project structure using namespaces
