+++ src/components/project-list.ts
@@ -1,0 +1,75 @@
+/// <reference path="base-component.ts" />
+
+namespace App {
+  // ProjectList Class
+  export class ProjectList
+    extends Component<HTMLDivElement, HTMLElement>
+    implements DragTarget {
+    assignedProjs: Project[];
+
+    constructor(private type: "active" | "finished") {
+      super("project-list", "app", false, `${type}-projects`);
+      this.assignedProjs = [];
+
+      this.configure();
+      this.renderContent();
+    }
+
+    @Autobind
+    dragOverHandler(event: DragEvent) {
+      if (event.dataTransfer && event.dataTransfer.types[0] === "text/plain") {
+        event.preventDefault();
+        const listEl = this.element.querySelector("ul")!;
+        listEl.classList.add("droppable");
+      }
+    }
+
+    @Autobind
+    dropHandler(event: DragEvent) {
+      const prjId = event.dataTransfer!.getData("text/plain");
+      const newStatus =
+        this.type === "active" ? ProjectStatus.Active : ProjectStatus.Finished;
+      projState.moveProject(prjId, newStatus);
+    }
+
+    @Autobind
+    dragLeaveHandler(_: DragEvent) {
+      const listEl = this.element.querySelector("ul")!;
+      listEl.classList.remove("droppable");
+    }
+
+    configure() {
+      this.element.addEventListener("dragover", this.dragOverHandler);
+      this.element.addEventListener("dragleave", this.dragLeaveHandler);
+      this.element.addEventListener("drop", this.dropHandler);
+
+      projState.addListener((projects: Project[]) => {
+        const relevantProjs = projects.filter((proj) => {
+          if (this.type === "active") {
+            return proj.status === ProjectStatus.Active;
+          }
+          return proj.status === ProjectStatus.Finished;
+        });
+        this.assignedProjs = relevantProjs;
+        this.renderProjects();
+      });
+    }
+
+    renderContent() {
+      const listId = `${this.type}-projects-list`;
+      this.element.querySelector("ul")!.id = listId;
+      this.element.querySelector("h2")!.textContent =
+        this.type.toUpperCase() + " PROJECTS";
+    }
+
+    private renderProjects() {
+      const listEl = document.getElementById(
+        `${this.type}-projects-list`
+      )! as HTMLUListElement;
+      listEl.innerHTML = "";
+      for (const prjItem of this.assignedProjs) {
+        new ProjectItem(this.element.querySelector("ul")!.id, prjItem);
+      }
+    }
+  }
+}
